\name{DALYmanual}

\alias{DALYmanual}

\title{Open the PDF \strong{DALY Calculator Manual}}

\description{
	This function opens the \strong{DALY Calculator Manual}, which is included as PDF file in the \strong{DALY} package.
}

\usage{DALYmanual()}

\author{Brecht Devleesschauwer, Arie Havelaar, Juanita Haagsma, Nicolas Praet, Niko Speybroeck}