\name{DALYoptions}

\alias{DALYoptions}

\title{Open the DALY Calculator options window}

\description{
	This function opens the \sQuote{options} window of the \emph{DALY Calculator}.
}

\usage{DALYoptions()}

\details{
	The following settings can be modified in the \sQuote{options} window:
	\itemize{
		\item{\strong{Iterations} the default number of iterations is 20,000}
		\item{\strong{Output} the basic output contains a summary of total DALYs, YLDs, YLLs and deaths, while the advanced output provides an additional outcome-wise summary; in addition, the user can specficy whether or not to view a standardized DALY histogram.}
	}
}

\author{Brecht Devleesschauwer, Arie Havelaar, Juanita Haagsma, Nicolas Praet, Niko Speybroeck}