\name{readDALYdata}

\alias{readDALYdata}

\title{Read population and epidemiological data from a tab-delimited text file}

\description{
	This function resets the \emph{DALY Calculator} (through \code{\link{reset}}), and then reads the population and epidemiological data from a tab-delimited text file that was created by the \emph{DALY Calculator} and possibly modified by the user.
}

\usage{readDALYdata()}

\author{Brecht Devleesschauwer, Arie Havelaar, Juanita Haagsma, Nicolas Praet, Niko Speybroeck}