\name{reset}
\alias{reset}

\title{Reset the DALY Calculator}

\description{
  This function resets the \emph{DALY Calculator} to its default values,
  *except* for the \sQuote{Life Expectancy} table and the options
  (cf \code{\link{DALYoptions}}).
}

\usage{reset()}

\author{
  Brecht Devleesschauwer,
  Arie Havelaar,
  Juanita Haagsma,
  Nicolas Praet,
  Niko Speybroeck

  \strong{Maintainer:} Brecht Devleesschauwer \email{brecht.devleesschauwer@ugent.be}
}

\seealso{
  \code{\link{DALYcalculator}}
  (for a brief description of the \emph{DALY Calculator})\cr
  \code{\link{DALYmanual}}
  (for a more comprehensive overview)
}
