\name{saveDALYdata}

\alias{saveDALYdata}

\title{Save population and epidemiological data to a tab-delimited text file}

\description{
	This function saves the population and epidemiological data to a tab-delimited text file, which can be opened in a spreadsheet program and modified by the user.
}

\usage{saveDALYdata()}

\author{Brecht Devleesschauwer, Arie Havelaar, Juanita Haagsma, Nicolas Praet, Niko Speybroeck}