\name{setDALYexample}

\alias{setDALYexample}

\title{Load one of the two built-in DALY calculation examples}

\description{
	This function resets the \emph{DALY Calculator} (through \code{\link{reset}}), and loads one of the two built-in DALY calculation examples:
	\enumerate{
		\item{Neurocysticercosis in West-Cameroon}
		\item{Congenital Toxoplasmosis in the Netherlands}
	}
}

\usage{setDALYexample(example)}

\arguments{
	\item{example}{This argument takes value \sQuote{\code{1}} for the Neurocysticercosis example, and value \sQuote{\code{2}} for the Toxoplasmosis example}
}

\author{Brecht Devleesschauwer, Arie Havelaar, Juanita Haagsma, Nicolas Praet, Niko Speybroeck}