\name{setData}

\alias{setData}

\title{Open a data input window}

\description{
	This function opens a data window where the input parameters (with corresponding distributions and stratification levels) for one of the disease categories or outcomes can be entered.
}

\usage{setData(n)}

\arguments{
	\item{n}{integer number, corresponding to the disease category or outcome to be set}
}

\author{Brecht Devleesschauwer, Arie Havelaar, Juanita Haagsma, Nicolas Praet, Niko Speybroeck}