\name{setLifeExp}

\alias{setLifeExp}

\title{Open the life expectancy window}

\description{
	This function opens the life expectancy window, where the life expectancy table can be customized. The default life expectancy table is the Coale and Demeny model life table West, which is set through function \code{\link{setStdLE}}.
}

\usage{setLifeExp}

\author{Brecht Devleesschauwer, Arie Havelaar, Juanita Haagsma, Nicolas Praet, Niko Speybroeck}