\name{setPop}
\alias{setPop}

\title{Open the \sQuote{Population} window}

\description{
  This function opens the \sQuote{Population} window,
  where the population under study can be entered, stratified by sex and age group.
}

\usage{setPop()}

\author{
  Brecht Devleesschauwer,
  Arie Havelaar,
  Juanita Haagsma,
  Nicolas Praet,
  Niko Speybroeck

  \strong{Maintainer:} Brecht Devleesschauwer \email{brecht.devleesschauwer@ugent.be}
}

\seealso{
  \code{\link{DALYcalculator}}
  (for a brief description of the \emph{DALY Calculator})\cr
  \code{\link{DALYmanual}}
  (for a more comprehensive overview)
}
