\name{setStdLE}

\alias{setStdLE}

\title{Set the life expectancy table to the standard life expectancy}

\description{
	This function sets the life expectancy table to the standard life expectancy for DALY calculation, i.e., the Coale and Demeny model life table West, level 25 and 26, which has a maximum life expectancy of 80 years for males and 82.5 years for females.
}

\usage{setStdLE}

\references{
	\itemize{
		\item{Murray CJL (1994). Quantifying the burden of disease: the technical basis for disability-adjusted life years. \emph{Bulletin of the World Health Organization 72(3),} 429-445.}
	}
}

\author{Brecht Devleesschauwer, Arie Havelaar, Juanita Haagsma, Nicolas Praet, Niko Speybroeck}