\name{DALYmanual}
\alias{DALYmanual}

\title{Open the PDF "\strong{DALY Calculator Manual}"}

\description{
	This function opens the "\strong{DALY Calculator Manual}",
	which is included as PDF file in the \strong{DALY} package.
}

\usage{DALYmanual()}

\author{\email{Brecht.Devleesschauwer@UGent.be}}

\seealso{
  \code{\link{DALYcalculator}}
  (for a brief description of the \emph{DALY Calculator})\cr
  \code{\link{DALYmanual}}
  (for a more comprehensive overview)
}
