\name{reset}
\alias{reset}

\title{Reset the DALY Calculator}

\description{
  This function resets the \emph{DALY Calculator} to its default values,
  *except* for the \sQuote{Life Expectancy} table and the options
  (cf \code{\link{DALYoptions}}).
}

\usage{reset()}

\author{\email{brechtdv@gmail.com}}

\seealso{
  \code{\link{DALYcalculator}}
  (for a brief description of the \emph{DALY Calculator})\cr
  \code{\link{DALYmanual}}
  (for a more comprehensive overview)
}
