\name{reset}

\alias{reset}

\title{Reset the DALY Calculator}

\description{
	This function resets the DALY Calculator to its default values, except for the life expectancy table.
}

\usage{reset()}

\author{Brecht Devleesschauwer, Arie Havelaar, Juanita Haagsma, Nicolas Praet, Niko Speybroeck}