\name{saveDALYdata}
\alias{saveDALYdata}

\title{Save population and epidemiological data to an .RData image file}

\description{
  This function saves the population and epidemiological data to an .RData image file.
  This file can be loaded into the \emph{DALY Calculator} through \code{readDALYdata()}.
}

\usage{saveDALYdata()}

\author{\email{brechtdv@gmail.com}}

\seealso{
  \code{\link{readDALYdata}}\cr
  
  \code{\link{DALYcalculator}}
  (for a brief description of the \emph{DALY Calculator})\cr
  \code{\link{DALYmanual}}
  (for a more comprehensive overview)
}

\keyword{import/export}
