\name{setData}
\alias{setData}

\title{Open a data input window}

\description{
  This function opens a data window where the input parameters
  (with corresponding distributions and stratification levels)
  for one of the disease categories or outcomes can be entered.
}

\usage{setData(n)}

\arguments{
  \item{n}{Integer, corresponding to the health outcome to be set}
}

\author{
  Brecht Devleesschauwer,
  Arie Havelaar,
  Juanita Haagsma,
  Nicolas Praet,
  Niko Speybroeck

  \strong{Maintainer:} Brecht Devleesschauwer \email{brecht.devleesschauwer@ugent.be}
}

\seealso{
  \code{\link{DALYcalculator}}
  (for a brief description of the \emph{DALY Calculator})\cr
  \code{\link{DALYmanual}}
  (for a more comprehensive overview)
}
