\name{setLifeExp}
\alias{setLifeExp}

\title{Open the \sQuote{Life Expectancy} window}

\description{
 This function opens the \sQuote{Life Expectancy} window,
 where the Life Expectancy table can be customized.
 The default table is the Coale and Demeny model life table West,
 which is set through function \code{\link{setStdLE}}.
}

\usage{setLifeExp()}

\author{\email{Brecht.Devleesschauwer@UGent.be}}

\seealso{
  \code{\link{setStdLE}}\cr\cr
  \code{\link{DALYcalculator}}
  (for a brief description of the \emph{DALY Calculator})\cr
  \code{\link{DALYmanual}}
  (for a more comprehensive overview)
}
