\name{setPop}
\alias{setPop}

\title{Open the \sQuote{Population} window}

\description{
  This function opens the \sQuote{Population} window,
  where the population under study can be entered, stratified by sex and age group.
}

\usage{setPop()}

\author{\email{Brecht.Devleesschauwer@UGent.be}}

\seealso{
  \code{\link{DALYcalculator}}
  (for a brief description of the \emph{DALY Calculator})\cr
  \code{\link{DALYmanual}}
  (for a more comprehensive overview)
}
