\name{setPop}

\alias{setPop}

\title{Open the population input window}

\description{
	This function opens the population window, where the population under study can be entered, stratified by sex and age group.
}

\usage{setPop()}

\author{Brecht Devleesschauwer, Arie Havelaar, Juanita Haagsma, Nicolas Praet, Niko Speybroeck}